MaxText/hope_alpha/requirements_for_wheel_downloads.txt
# Before being renamed (from requirements.txt) this file was used to download pip wheels using the following command: 
#pip download -r requirements.txt -d /mnt/arcus/lab/shared/maxtext_pip_wheels/maxtext_pip_wheels_main

#--find-links /home/sa_112413943341569390945/llama3.1-whls

jax>=0.4.30
jaxlib>=0.4.30
orbax-checkpoint>=0.5.21
absl-py
array-record
aqtp
cloud-accelerator-diagnostics
cloud-tpu-diagnostics
datasets
gcsfs
google-cloud-aiplatform==1.61.0
google-cloud-storage
grain-nightly
flax>=0.8.0
ml-collections
ml-goodput-measurement
numpy
optax
protobuf==3.20.3
pylint
pytest
pyink
pre-commit
pytype
sentencepiece @ https://files.pythonhosted.org/packages/ec/87/f26695307c0aa00e6938f5de795fc7f2c718a448b48d29a4c8c8dbf829d3/sentencepiece-0.1.97.tar.gz
tensorflow-text>=2.13.0
tensorflow>=2.13.0
tensorflow-datasets
tensorboardx
tensorboard-plugin-profile
tiktoken
transformers
#mlperf-logging
mlperf-logging @ git+https://github.com/mlperf/logging.git
google-jetstream
jsonlines
